+++ bin/index.php
@@ -12,18 +12,12 @@ use CastroItalo\EchoQuery\Builder;
  */
 function main(): void
 {
-    $sub_query = (new Builder())->select(
-        ['column_one'],
-        ['column_two']
-    )
-        ->from('table_one', 'to')
-        ->__toString();
-
     $query = (new Builder())->select(
         ['a.column_one', 'co'],
         ['a.column_two', 'ct']
     )
-        ->from($sub_query, 'a', true)
+        ->from('table_name', 'a')
+        ->where('a.column_one')
         ->__toString();
 
     echo $query . PHP_EOL;