+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 
 All notable changes to `EchoQuery` will be documented in this file.
 
-## Last update 03-04-2024
+## Last update 04-04-2024
 
 ### Added
-- Sub query on FROM statement
+- Base where statement done and tested
 
 ### Deprecated
 - Nothing